Updated extractor to use EternalREsourceExtractor; Updated readme

+++ EternalExtractor/EternalExtractor.sh
@@ -16,6 +16,9 @@
 # You should have received a copy of the GNU General Public License  
 # along with EternalBasher. If not, see <https://www.gnu.org/licenses/>.
 
+
+# Functions
+
 MissingDoomEternal(){
 	printf "'DOOMEternalx64vk.exe' not found!
 	
@@ -34,9 +37,16 @@ MissingResources(){
 }
 
 MissingQuickBMS(){
-	printf "quickbms not found!
+	printf "QuickBMS not found!
 	Is your QuickBMS installation incomplete, or did you use a wrong path?
-	The path that you gave was %s" "$___QUICKBMS_DIR"
+	The path that you gave was %s" "$___EXTRACTOR_DIR"
+	exit
+}
+
+MissingEternalResourceExtractor(){
+	printf "EternalResourceExtractor not found!
+	Did you make sure to download the executable or did you use a wrong path?
+	The path that you gave was %s" "$___EXTRACTOR_DIR"
 	exit
 }
 
@@ -54,22 +64,62 @@ OutputIsntEmpty(){
 	exit
 }
 
+# Ask for QuickBMS path
+
+QuickBMSPath(){
+	printf "How did you install QuickBMS?\n
+			1 - Compiled the source code with 'make'/Downloaded the Linux standalone binaries.\n
+			2 - Installed from the Arch User Repository(AUR)\n
+			3 - WINE version\n
+			4 - Exit\n\n"
+	
+	while true
+	do
+		read -r ___QUICKBMS_SOURCE
+		case "$___QUICKBMS_SOURCE" in
+			"1")
+				read -rp "Please type the path to your QuickBMS directory: " inp
+				___EXTRACTOR_DIR="$inp"
+				break
+				;;
+			"2")
+				___EXTRACTOR_DIR="/usr/bin/"
+				break
+				;;
+			"3")
+				read -rp "Please type the path to your QuickBMS directory: " inp
+				___EXTRACTOR_DIR="$inp"
+				break
+				;;
+			"4")
+				echo "Exiting script."
+				exit 1
+				;;
+			*)
+				read -rp "Not a valid option. Please try again." ___QUICKBMS_SOURCE
+				;;
+		esac
+	done
+
+	if ! [[ "$___EXTRACTOR_DIR" == */ ]]; then ___EXTRACTOR_DIR="${___EXTRACTOR_DIR}/"; fi
+}
+
+# Start script
+
 printf "Eternal Extractor Bash script\n
 			by Leveste and PowerBall253\n
 			based on the original file by Zwip-Zwap Zapony\n\n\n"
 
 export ___GAMEDIR="$HOME/.local/share/Steam/steamapps/common/DOOMEternal/"
 export ___OUTPUT_DIR=""
-export ___QUICKBMS_DIR=""
+export ___EXTRACTOR_DIR=""
 export ___QUICKBMS_SCRIPT=""
 
 
 
 
-printf "This batch file runs QuickBMS to extract the contents of all of DOOM Eternal's *.resources archives in one go.\n\n
+printf "This shell script will help run QuickBMS or EternalResourceExtractor to extract the contents of all of DOOM Eternal's *.resources archives in one go.\n\n
 If you only want to use others' DOOM Eternal mods, not make your own, this isn't useful for you; simply close this window now.\n\n
-Otherwise, please input the full filepath to your DOOM Eternal installation:\n\n
-			This script assumes that you have placed the 'doometernal.bms' file in your game directory. If you haven't, please move or copy it over.\n\n
 			The default location for your game directory is set to %s.\n
 			Press 'Y' if you wish to change it. Press any other key to continue with this setting." "$___GAMEDIR"
 
@@ -85,49 +135,6 @@ if ! [[ "$___GAMEDIR" == */ ]]; then ___GAMEDIR="${___GAMEDIR}/"; fi
 if ! [ -f "${___GAMEDIR}DOOMEternalx64vk.exe" ]; then MissingDoomEternal; fi
 if ! [ -f "${___GAMEDIR}base/gameresources.resources" ]; then MissingResources; fi
 
-#Ask for QuickBMS path
-
-printf "How did you install QuickBMS?\n
-		1 - Compiled the source code with 'make'/Downloaded the Linux standalone binaries.\n
-		2 - Installed from the Arch User Repository(AUR)\n
-		3 - WINE version\n
-		4 - Exit\n\n"
-
-while true
-do
-	read -r ___QUICKBMS_SOURCE
-	case "$___QUICKBMS_SOURCE" in
-		"1")
-			read -rp "Please type the path to your QuickBMS directory: " inp
-			___QUICKBMS_DIR="$inp"
-			break
-			;;
-		"2")
-			___QUICKBMS_DIR="/usr/bin/"
-			break
-			;;
-		"3")
-			read -rp "Please type the path to your QuickBMS directory: " inp
-			___QUICKBMS_DIR="$inp"
-			break
-			;;
-		"4")
-			echo "Exiting script."
-			exit 1
-			;;
-		*)
-			read -rp "Not a valid option. Please try again." ___QUICKBMS_SOURCE
-			;;
-	esac
-done
-if ! [[ "$___QUICKBMS_DIR" == */ ]]; then ___QUICKBMS_DIR="${___QUICKBMS_DIR}/"; fi
-
-if ! ([ -f "${___QUICKBMS_DIR}quickbms" ] || [ -f "${___QUICKBMS_DIR}quickbms_4gb_files.exe" ]); then MissingQuickBMS; fi
-if [ -f "${___GAMEDIR}doometernal.txt" ]; then ___QUICKBMS_SCRIPT="${___GAMEDIR}doometernal.txt"; fi
-if [ -f "${___GAMEDIR}doometernal.bms.txt" ]; then ___QUICKBMS_SCRIPT="${___GAMEDIR}doometernal.bms.txt"; fi
-if [ -f "${___GAMEDIR}doometernal.bms" ]; then ___QUICKBMS_SCRIPT="${___GAMEDIR}doometernal.bms"; fi
-if [ -z ${___QUICKBMS_SCRIPT+x} ]; then MissingScript; fi
-
 #Ask for output path
 
 printf "Please input the full filepath to where you want to extract resources to.
@@ -141,30 +148,69 @@ if [ -n "$(ls -A "$___OUTPUT_DIR")" ]
 then  OutputIsntEmpty
 fi
 
-printf "The expected filesize required to extract DOOM Eternal v3.1's resources is 12.2 gigabytes. (Expect more in later updates.)\n
-	Please note that extracting all *.resources archives might take hours, depending on your CPU and storage speed!\n\n
-	And, just to make sure, does this look correct?
 
-	DOOM Eternal: %s
-	QuickBMS: %s
-	Output: %s
-	\n\n
-	Press any key to continue. To exit the script, use Ctrl-C to cancel." "$___GAMEDIR" "$___QUICKBMS_DIR" "$___OUTPUT_DIR"
+# Ask for which extractor tool to use
 
-read -n 1 -s # Wait for user keypress
-
-#Prompt to start extraction
+printf "Which extractor are you using? \n
+		1 - QuickBMS\n
+		2 - EternalResourceExtractor\n\n"
 
 
-case "$___QUICKBMS_SOURCE" in
-	[12])
-		print "Linux version of quickbms does not work with the current 'doometernal.bms' script due to the 'oodle' library's open source alternative not being implemented yet. Please use the WINE version instead. The script will be updated when it's been implemented."
-		#cp "${___QUICKBMS_DIR}quickbms" .
-		#find "$___GAMEDIR" -name "*.resources" -exec sh -c './quickbms -o -Y "$___QUICKBMS_SCRIPT" "$1" "$___OUTPUT_DIR"' sh {} \;
-		#rm quickbms
-		;;
-	"3")
-		cp "${___QUICKBMS_DIR}quickbms_4gb_files.exe" .
-		find "$___GAMEDIR" -name "*.resources" -exec sh -c 'wine quickbms_4gb_files.exe -o -Y "$___QUICKBMS_SCRIPT" "$1" "$___OUTPUT_DIR"' sh {} \;
-		rm quickbms_4gb_files.exe
-esac
+while true
+do
+	read -r inp
+	case "$inp" in
+		"1")
+			QuickBMSPath
+			
+			if ! ([ -f "${___EXTRACTOR_DIR}quickbms" ] || [ -f "${___EXTRACTOR_DIR}quickbms_4gb_files.exe" ]); then MissingQuickBMS; fi
+			if [ -f "${___GAMEDIR}doometernal.txt" ]; then ___QUICKBMS_SCRIPT="${___GAMEDIR}doometernal.txt"; fi
+			if [ -f "${___GAMEDIR}doometernal.bms.txt" ]; then ___QUICKBMS_SCRIPT="${___GAMEDIR}doometernal.bms.txt"; fi
+			if [ -f "${___GAMEDIR}doometernal.bms" ]; then ___QUICKBMS_SCRIPT="${___GAMEDIR}doometernal.bms"; fi
+			if [ -z ${___QUICKBMS_SCRIPT+x} ]; then MissingScript; fi
+			#Prompt to start extraction
+			
+			
+			case "$___QUICKBMS_SOURCE" in
+				[12])
+					print "Linux version of quickbms does not work with the current 'doometernal.bms' script due to the 'oodle' library's open source alternative not being implemented yet. Please use the WINE version instead. The script will be updated when it's been implemented."
+					#cp "${___EXTRACTOR_DIR}quickbms" .
+					#find "$___GAMEDIR" -name "*.resources" -exec sh -c './quickbms -o -Y "$___QUICKBMS_SCRIPT" "$1" "$___OUTPUT_DIR"' sh {} \;
+					#rm quickbms
+					;;
+				"3")
+					if [ -f "quickbms_4gb_files.exe" ]
+					then
+						find "$___GAMEDIR" -name "*.resources" -exec sh -c 'wine quickbms_4gb_files.exe -o -Y "$___QUICKBMS_SCRIPT" "$1" "$___OUTPUT_DIR"' sh {} \;
+					else
+						cp "${___EXTRACTOR_DIR}quickbms_4gb_files.exe" .
+						find "$___GAMEDIR" -name "*.resources" -exec sh -c 'wine quickbms_4gb_files.exe -o -Y "$___QUICKBMS_SCRIPT" "$1" "$___OUTPUT_DIR"' sh {} \;
+						rm quickbms_4gb_files.exe
+					fi
+			esac
+
+			break
+			;;
+		"2")
+			if [ -f "EternalResourceExtractor" ]
+			then
+				find "$___GAMEDIR" -name "*.resources" -exec sh -c './EternalResourceExtractor "$1" "$___OUTPUT_DIR"' sh {} \;
+			else
+				read -rp "Please type the path to your EternalResourceExtractor directory: " inp
+				___EXTRACTOR_DIR="$inp"
+				if ! [[ "$___EXTRACTOR_DIR" == */ ]]; then ___EXTRACTOR_DIR="${___EXTRACTOR_DIR}/"; fi
+
+				if ! ([ -f "${___EXTRACTOR_DIR}EternalResourceExtractor" ]); then MissingEternalResourceExtractor; fi
+
+				cp "${___EXTRACTOR_DIR}EternalResourceExtractor" .
+				chmod +x "EternalResourceExtractor"
+				find "$___GAMEDIR" -name "*.resources" -exec sh -c './EternalResourceExtractor "$1" "$___OUTPUT_DIR"' sh {} \;
+				rm EternalResourceExtractor
+			fi
+			break
+			;;
+		"*")
+			read -rp "Not a valid option. Please try again." inp
+			;;
+	esac
+done
